Add jwtDecode and resolve refresh token when signing with provider problems
