+++ .env.local.example
@@ -27,5 +27,5 @@ FEATURE_CAMPAIGN=true
 ##############
 NEXTAUTH_URL=http://localhost:3040
 NEXTAUTH_SECRET=super-dev-secret
-GOOGLE_ID=
-GOOGLE_SECRET=
+GOOGLE_ID=333154281816-5rfm87qo5s98ubslj44fo9o2h2636sdd.apps.googleusercontent.com
+GOOGLE_SECRET=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
