+++ package.json
@@ -52,6 +52,7 @@
     "formik": "2.2.9",
     "highcharts": "9.1.0",
     "highcharts-react-official": "3.0.0",
+    "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "mobx": "6.3.2",
     "mobx-react": "7.2.0",
