--- src/common/util/parseJWT.ts
@@ -1,7 +1,0 @@
-export function parseJWT<T>(token: string): T | null {
-  try {
-    return JSON.parse(atob(token.split('.')[1]))
-  } catch (e) {
-    return null
-  }
-}
